Make the transformer tests test the code
Two sampler tests re-implemented the batch-wrapping branch inline and never
called _run_sampler; one parametrized test now calls it with stubbed transform
and network and checks the shapes handed to the network and squeezed back out.
Two uniformity assertions used atol=5, rtol=5, which accept any value; they are
chi-square tests now, seeded. Delete a test of torch's own LayerNorm, a
duplicate embedding-type test and a shape test subsumed by its neighbour. Move
the detect_asd_notches tests next to the function, in test_gwutils.py, and fix
the bin count in a fixture comment (8033, not 8032).
