+++ tests/gw/test_gwutils.py
@@ -1,14 +1,18 @@
+import warnings
+
 import numpy as np
 import pandas as pd
 import pytest
 
-from dingo.gw.domains import UniformFrequencyDomain
+from dingo.gw.domains import UniformFrequencyDomain, MultibandedFrequencyDomain
 from dingo.gw.gwutils import (
+    detect_asd_notches,
     get_extrinsic_prior_dict,
     get_mismatch,
     get_standardization_dict,
     get_window,
 )
+from dingo.gw.noise.asd_dataset import HIGH_ASD_VALUE
 
 
 # ---------------------------------------------------------------------------
@@ -148,3 +152,108 @@ def test_get_standardization_dict_rejects_nonzero_intrinsic_std_for_extrinsic(
     wfd = _StubWaveformDataset(luminosity_distance_std=5.0)
     with pytest.raises(ValueError, match="fixed value"):
         get_standardization_dict(extrinsic_prior, wfd, ["chirp_mass"])
+
+
+# ---------------------------------------------------------------------------
+# detect_asd_notches
+# ---------------------------------------------------------------------------
+
+
+def _make_asd_array(domain, notch_intervals=None):
+    """Build a full-length ASD array (length max_idx + 1) with real-valued noise
+    below HIGH_ASD_VALUE, except inside notch_intervals where ASD = HIGH_ASD_VALUE.
+    Edge-padding bins (0 .. min_idx-1) are set to HIGH_ASD_VALUE by convention.
+    """
+    n = domain.max_idx + 1
+    asd = np.full(n, 1e-23)  # realistic noise amplitude
+    # edge padding
+    asd[: domain.min_idx] = HIGH_ASD_VALUE
+    if notch_intervals:
+        freqs = domain.sample_frequencies
+        for f_lo, f_hi in notch_intervals:
+            mask = (freqs >= f_lo) & (freqs <= f_hi)
+            asd[mask] = HIGH_ASD_VALUE
+    return asd
+
+
+def test_detect_asd_notches_no_notch():
+    """No notches in any detector → returns None."""
+    domain = UniformFrequencyDomain(f_min=20.0, f_max=512.0, delta_f=0.125)
+    asd = _make_asd_array(domain)
+    result = detect_asd_notches({"H1": asd, "L1": asd}, domain)
+    assert result is None
+
+
+def test_detect_asd_notches_single_notch():
+    """Single interior notch is correctly detected."""
+    domain = UniformFrequencyDomain(f_min=20.0, f_max=512.0, delta_f=0.125)
+    f_lo, f_hi = 59.0, 61.0
+    asd = _make_asd_array(domain, notch_intervals=[[f_lo, f_hi]])
+    result = detect_asd_notches({"H1": asd}, domain)
+    assert result is not None
+    assert "H1" in result
+    intervals = result["H1"]
+    assert len(intervals) == 1
+    detected_lo, detected_hi = intervals[0]
+    assert detected_lo >= f_lo
+    assert detected_hi <= f_hi + domain.delta_f
+
+
+def test_detect_asd_notches_multiple_notches():
+    """Multiple disjoint notches per detector are all detected."""
+    domain = UniformFrequencyDomain(f_min=20.0, f_max=512.0, delta_f=0.125)
+    notches = [[59.0, 61.0], [119.0, 121.0]]
+    asd = _make_asd_array(domain, notch_intervals=notches)
+    result = detect_asd_notches({"H1": asd}, domain)
+    assert result is not None
+    assert len(result["H1"]) == 2
+
+
+def test_detect_asd_notches_edge_padding_ignored():
+    """Edge-padding at f_min (index 0 of valid band) is not reported as a notch."""
+    domain = UniformFrequencyDomain(f_min=20.0, f_max=512.0, delta_f=0.125)
+    # Set the very first valid bin (at f_min) to HIGH_ASD_VALUE to mimic edge padding.
+    asd = _make_asd_array(domain)
+    asd[domain.min_idx] = HIGH_ASD_VALUE
+    result = detect_asd_notches({"H1": asd}, domain)
+    assert result is None
+
+
+def test_detect_asd_notches_edge_padding_ignored_at_f_max():
+    """A high run touching f_max (bilby fills frequencies beyond a PSD file with inf)
+    is edge padding, not a notch. One bin is silent; wider warns that the PSD does
+    not cover the model band."""
+    domain = UniformFrequencyDomain(f_min=20.0, f_max=512.0, delta_f=0.125)
+    asd = _make_asd_array(domain)
+    asd[-1] = np.inf
+    with warnings.catch_warnings():
+        warnings.simplefilter("error")
+        assert detect_asd_notches({"H1": asd}, domain) is None
+    asd[-3:] = np.inf
+    with pytest.warns(UserWarning, match="does not cover"):
+        assert detect_asd_notches({"H1": asd}, domain) is None
+
+
+def test_detect_asd_notches_per_detector():
+    """Notch in H1 only is not reported for L1."""
+    domain = UniformFrequencyDomain(f_min=20.0, f_max=512.0, delta_f=0.125)
+    asd_h1 = _make_asd_array(domain, notch_intervals=[[59.0, 61.0]])
+    asd_l1 = _make_asd_array(domain)
+    result = detect_asd_notches({"H1": asd_h1, "L1": asd_l1}, domain)
+    assert result is not None
+    assert "H1" in result
+    assert "L1" not in result
+
+
+def test_detect_asd_notches_multibanded_uses_base_domain():
+    """Stored ASDs live on the base grid; MFD indices must not be used (this
+    returned wrong notch frequencies before)."""
+    base = UniformFrequencyDomain(f_min=20.0, f_max=100.0, delta_f=0.25)
+    mfd = MultibandedFrequencyDomain(
+        nodes=[20.0, 36.0, 100.0], delta_f_initial=0.25, base_domain=base
+    )
+    asd = np.full(len(base), 1e-23)
+    freqs = base.sample_frequencies
+    asd[(freqs >= 60.0) & (freqs <= 61.0)] = HIGH_ASD_VALUE
+    notches = detect_asd_notches({"H1": asd}, mfd)
+    assert notches == {"H1": [[60.0, 61.0]]}
